refactor: share QA runtime helpers
