+++ extensions/qa-lab/api.ts
@@ -47,7 +47,11 @@ export {
   QA_BASE_RUNTIME_PLUGIN_IDS,
   type QaThinkingLevel,
 } from "./src/qa-gateway-config.js";
-export { type QaReportCheck, type QaReportScenario, renderQaMarkdownReport } from "./src/report.js";
+export {
+  renderQaMarkdownReport,
+  type QaReportCheck,
+  type QaReportScenario,
+} from "openclaw/plugin-sdk/qa-runner-runtime";
 export {
   type QaScenarioDefinition,
   type QaScenarioResult,
