+++ extensions/qa-lab/src/live-transports/discord/cli.runtime.ts
@@ -1,8 +1,6 @@
+import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runner-runtime";
 import type { LiveTransportQaCommandOptions } from "../shared/live-transport-cli.js";
-import {
-  printLiveTransportQaArtifacts,
-  resolveLiveTransportQaRunOptions,
-} from "../shared/live-transport-cli.runtime.js";
+import { resolveLiveTransportQaRunOptions } from "../shared/live-transport-cli.runtime.js";
 import { runDiscordQaLive } from "./discord-live.runtime.js";
 
 export async function runQaDiscordCommand(opts: LiveTransportQaCommandOptions) {
