+++ extensions/qa-lab/src/live-transports/shared/live-gateway.runtime.ts
@@ -1,4 +1,5 @@
 import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
+import { appendQaLiveLaneIssue as appendLiveLaneIssue } from "openclaw/plugin-sdk/qa-runner-runtime";
 import {
   startQaGatewayChild,
   type QaCliBackendAuthMode,
@@ -7,7 +8,6 @@ import {
 import type { QaProviderMode } from "../../model-selection.js";
 import { startQaProviderServer } from "../../providers/server-runtime.js";
 import type { QaThinkingLevel } from "../../qa-gateway-config.js";
-import { appendLiveLaneIssue } from "./live-lane-helpers.js";
 
 async function stopQaLiveLaneResources(
   resources: {
