+++ extensions/qa-lab/src/live-transports/shared/live-transport-cli.runtime.ts
@@ -37,12 +37,3 @@ export function resolveLiveTransportQaRunOptions(
     credentialRole: opts.credentialRole?.trim(),
   };
 }
-
-export function printLiveTransportQaArtifacts(
-  laneLabel: string,
-  artifacts: Record<string, string>,
-) {
-  for (const [label, filePath] of Object.entries(artifacts)) {
-    process.stdout.write(`${laneLabel} ${label}: ${filePath}\n`);
-  }
-}
