+++ extensions/qa-matrix/src/runners/contract/runtime.ts
@@ -4,14 +4,14 @@ import path from "node:path";
 import { setTimeout as sleep } from "node:timers/promises";
 import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
 import { formatErrorMessage } from "openclaw/plugin-sdk/error-runtime";
-import { loadQaRuntimeModule } from "openclaw/plugin-sdk/qa-runner-runtime";
-import type { QaReportCheck } from "../../report.js";
-import { renderQaMarkdownReport } from "../../report.js";
-import { normalizeQaProviderMode, type QaProviderModeInput } from "../../run-config.js";
 import {
-  appendLiveLaneIssue,
-  buildLiveLaneArtifactsError,
-} from "../../shared/live-lane-helpers.js";
+  appendQaLiveLaneIssue as appendLiveLaneIssue,
+  buildQaLiveLaneArtifactsError as buildLiveLaneArtifactsError,
+  loadQaRuntimeModule,
+  renderQaMarkdownReport,
+  type QaReportCheck,
+} from "openclaw/plugin-sdk/qa-runner-runtime";
+import { normalizeQaProviderMode, type QaProviderModeInput } from "../../run-config.js";
 import { buildMatrixQaObservedEventsArtifact } from "../../substrate/artifacts.js";
 import { provisionMatrixQaRoom, type MatrixQaProvisionResult } from "../../substrate/client.js";
 import {
