+++ extensions/qa-matrix/src/shared/live-transport-cli.runtime.test.ts
@@ -1,8 +1,8 @@
 import { mkdir, mkdtemp, readFile, rm } from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
+import { startLiveTransportQaOutputTee } from "openclaw/plugin-sdk/qa-runner-runtime";
 import { afterEach, describe, expect, it } from "vitest";
-import { startLiveTransportQaOutputTee } from "./live-transport-cli.runtime.js";
 
 const tmpDirs: string[] = [];
